Fix interpretation of 3DS statuses
The previous implementation of the ServerRestCompleteResponse would have
suggested that a transaction was fully authorised if the 3DS status was
authenticated even if the transaction itself was declined by the bank,
because the transaction status was only checked as a fallback if the 3DS
status was not authenticated.

Now, both the transaction and 3DS statuses must be "successful".